\name{get.similarity}
\alias{get.similarity}
\title{
Interogate similarity function for Geary-Moran TDT tests 
}
\usage{
get.similarity(nl=1)
}
\arguments{
\item{nl}{
The number of loci.
}}
\description{
N/A
}
\seealso{
\code{\link{tdt.quad}}, \code{\link{set.similarity}}
}
\keyword{htest}
% Converted by Sd2Rd version 0.2-a4.
